Separated Auth endpoints from user endpoints, adjusted front to new controller url base
